📝 Better swagger documentation
