+++ src/api/module/community/communityModule.ts
@@ -1,4 +1,5 @@
-import { CommunityDto, CommunityDtoMeta } from '@contracts/dto/forms/communityDto';
+import { CommunityDto } from '@contracts/dto/forms/communityDto';
+import { CommunityDtoMeta } from '@contracts/dto/forms/meta/communityDto.meta';
 import { cyrb53 } from '@helpers/hashHelper';
 import { getDatabaseService } from '@services/external/database/databaseService';
 import { Community } from '@services/external/database/xata';
