+++ src/api/swagger/registerSwaggerModuleComponents.ts
@@ -7,14 +7,50 @@ export const registerSwaggerModuleComponents = (props: { swaggerBuilder: Swagger
     const module = moduleLookup[segment as keyof IApiSegment];
 
     let propsObj = {};
-    for (const dtoMeta of Object.keys(module.dtoMeta)) {
+    for (const dtoMetaKey of Object.keys(module.dtoMeta)) {
+      const dtoMetaObj = module.dtoMeta[dtoMetaKey];
+      let typeDescrip: Record<string, any> | null = {
+        type: 'string',
+      };
+      if (dtoMetaObj.swaggerSchema != null) {
+        typeDescrip = {};
+        if (dtoMetaObj.swaggerSchema?.ref != null) {
+          typeDescrip.$ref = dtoMetaObj.swaggerSchema.ref;
+        }
+        if (dtoMetaObj.swaggerSchema?.type != null) {
+          typeDescrip.type = dtoMetaObj.swaggerSchema.type;
+        }
+        if (dtoMetaObj.swaggerSchema?.format != null) {
+          typeDescrip.format = dtoMetaObj.swaggerSchema.format;
+        }
+        const propsThatNeedItemProp = [
+          dtoMetaObj.swaggerSchema?.itemsType,
+          dtoMetaObj.swaggerSchema?.itemsRef,
+          dtoMetaObj.swaggerSchema?.itemsFormat,
+        ];
+        if (propsThatNeedItemProp.filter(Boolean).length > 0) {
+          typeDescrip.items = {};
+        }
+        if (dtoMetaObj.swaggerSchema?.itemsType != null) {
+          typeDescrip.items.type = dtoMetaObj.swaggerSchema.itemsType;
+        }
+        if (dtoMetaObj.swaggerSchema?.itemsRef != null) {
+          typeDescrip.items.$ref = dtoMetaObj.swaggerSchema.itemsRef;
+        }
+        if (dtoMetaObj.swaggerSchema?.itemsFormat != null) {
+          typeDescrip.items.format = dtoMetaObj.swaggerSchema.itemsFormat;
+        }
+      }
+
       propsObj = {
         ...propsObj,
-        [dtoMeta]: {
+        [dtoMetaKey]: {
+          ...typeDescrip,
           nullable: false,
         },
       };
     }
+
     props.swaggerBuilder.addComponent({
       [module.name]: {
         type: 'object',
